move plants tests files into corresponding folders
