+++ tests/repository_tests/test_plants_repository.py
